+++ src/extra_checks/__init__.py
@@ -15,6 +15,8 @@ class CheckId(str, enum.Enum):
     X056 = "field-boolean-null"
     X057 = "field-null"
     X058 = "field-foreign-key-db-index"
+    X301 = "drf-model-serializer-extra-kwargs"
+    X302 = "drf-model-serializer-meta-attribute"
 
 
 _IGNORED = {}