+++ src/extra_checks/checks/__init__.py
@@ -1,3 +1,8 @@
 from .model_checks import *  # noqa
 from .model_field_checks import *  # noqa
 from .self_checks import *  # noqa
+
+try:
+    from .drf_serializer_checks import *  # noqa
+except ImportError:
+    pass